requirements.txt
autopep8==2.0.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
mysql-connector-python==8.0.31
protobuf==3.20.1
pycodestyle==2.10.0
pymssql==2.2.7
pyodbc==4.0.35
requests==2.28.1
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.13
